Change of IIR structure
Introduction of ASTNodeIterator, StatementAccessesPairIterator. BlockStatements removed (still getters in SAP class are provided to keep compatibility with some parts of the code). StatementAccessesPairs are not IIRNodes anymore, but values in the ASTStmtToSAPMap of  a DoMethod. All uses of doMethod.getChildren() / iterateIIROver<SAP> / sap.getBlockStatements()  are substituted with method calls that return suitable iterators/ranges. DoMethod now applies the StatementMapper on itself and calculate the accesses of its statements inside its constructor (for a stencil) and provides appendStatement() / appendStatements() methods that also keep it in a consistent state.